+++ include/Operators/IOperator.hpp
@@ -15,8 +15,8 @@ private:
 public:
   virtual ~IOperator(){};
 
-  virtual double compute(int position, const DataSet &data,
-                        const std::vector<int> &index, const double criteria) const = 0;
+  virtual double compute(size_t position, const DataSet &data,
+                        const std::vector<size_t> &index, const double criteria) const = 0;
 
   virtual void print() = 0;
 };
