+++ include/PredictionModel/Trees/TrainingElement.hpp
@@ -12,15 +12,16 @@ struct TrainingElement {
 private:
   // Parameters
 
-  int depth;
+  uint16_t depth;
   TreeNode *node;
-  std::vector<int> index;
+  std::vector<size_t> index;
 
 public:
   // Constructor
 
   TrainingElement();
-  TrainingElement(TreeNode *node, const std::vector<int> &index, int depth);
+  TrainingElement(TreeNode *node, const std::vector<size_t> &index,
+                  uint16_t depth);
   TrainingElement(const TrainingElement &TE);
 
   // Operator Overloading
@@ -33,18 +34,18 @@ public:
 
   // Getters
 
-  const std::vector<int> &get_Index();
+  const std::vector<size_t> &get_Index();
 
   // Setters
-  void set_depth(int depth);
+  void set_depth(uint16_t depth);
   void set_Node(TreeNode *node);
-  void set_Index(const std::vector<int> &index);
+  void set_Index(const std::vector<size_t> &index);
 
   // Methods
 
   static void train(const DataSet &data, TreeNode *node,
                     const IOperator *splitting_Operator,
-                    const ICriteria *splitting_Criteria, int max_Depth,
+                    const ICriteria *splitting_Criteria, uint16_t max_Depth,
                     size_t treshold);
 
 private:
@@ -52,20 +53,22 @@ private:
   split_Node(const DataSet &data, const IOperator *splitting_Operator,
              const ICriteria *splitting_Criteria);
 
-  std::tuple<int, double> find_Best_Split(const DataSet &data,
-                                          const IOperator *splitting_Operator,
-                                          const ICriteria *splitting_Criteria);
+  std::tuple<size_t, double>
+  find_Best_Split(const DataSet &data, const IOperator *splitting_Operator,
+                  const ICriteria *splitting_Criteria);
 
-  std::tuple<std::optional<std::vector<int>>, std::optional<std::vector<int>>>
-  split_Index(const DataSet &data, int criterion, int position);
+  std::tuple<std::optional<std::vector<size_t>>,
+             std::optional<std::vector<size_t>>>
+  split_Index(const DataSet &data, double criterion, size_t position);
 
-  void set_Root(int dataset_Size, TreeNode *node);
-  void bootstrap_Index(int dataset_Size);
+  void set_Root(size_t dataset_Size, TreeNode *node);
+  void bootstrap_Index(size_t dataset_Size);
 
   /*****************************************************************************/
-  std::tuple<int, double> find_Best_Split_Parallel(const DataSet &data,
-                                          const IOperator *splitting_Operator,
-                                          const ICriteria *splitting_Criteria);
+  std::tuple<size_t, double>
+  find_Best_Split_Parallel(const DataSet &data,
+                           const IOperator *splitting_Operator,
+                           const ICriteria *splitting_Criteria);
 };
 
 #endif
